moved schedule and queue pages
